ReadString: Consume DecodingInputStream end markers.
org.freehep.postscript.operators.FileOperator (ReadString.execute): If
(and only if) the underlying InputStream is a DecodingInputStream,
consume all its input even if there is some remaining after our string
is already full.

This fixes use of DecodingInputStreams, where each DecodingInputStream
has its own end marker. If we just stop reading after our string is
full, the underlying DecodingInputStream will never consume its end
marker, and that end marker will be left in the underlying file's
InputStream. Hence, the next read from that underlying file will then
return that end marker and we will read a bogus empty string. This fix
works around that.

There may or may not be a better way to fix this, but this works for
now.