PSGState: Do not initialize paint to null.
org.freehep.postscript.types.PSGState (constructor): Initialize paint to
the underlying device's current paint, not to null.
(apply): Remove the paint != null check, paint should never be null.

This fixes grestore to the default color. (It did not appear in my test
cases, but it is possible, and grestore would not have done anything to
the color in that case without this fix.)